+++ but-what-about/CHANGELOG.md
@@ -1,3 +1,9 @@
 ### Unreleased
 
+#### New Features
+
+* Added stairs, slabs, and walls for Calcite, Cracked Deepslate Bricks, Cracked Deepslate Tiles,
+  Cracked Nether Bricks, Cracked Polished Blackstone Bricks, Cracked Stone Bricks, End Stone,
+  Netherrack, and Smooth Basalt.
+
 #### Changes